requirements.txt
openai>=1.12.0
pydantic>=2.0
python-dotenv>=1.0.0
numpy>=1.24.0
edge-tts>=6.1.0
streamlit>=1.30.0
pytest>=7.0
